refactor: 사용하지 않는 파일 삭제 및 분산된 파일들 합침
