--- TVApp/TVApp/Model/Channel.swift
@@ -1,14 +1,0 @@
-//
-//  Channel.swift
-//  TVApp
-//
-//  Created by 지현우 on 2021/01/26.
-//
-
-import Foundation
-
-struct Channel: Codable {
-    var name: String
-    var visitCount: Int
-    var createTime: String
-}
