+++ TVApp/TVApp/Protocols/DiffType.swift
@@ -7,7 +7,28 @@
 
 import Foundation
 
-enum Section: Hashable{
+enum Section: Hashable {
     case main
 }
 
+struct Item: Hashable {
+    // MARK: Lifecycle
+
+    init(_ data: DataType) {
+        self.data = data
+        self.id = UUID()
+    }
+
+    // MARK: Internal
+
+    let id: UUID
+    let data: DataType
+
+    static func == (lhs: Item, rhs: Item) -> Bool {
+        lhs.id == rhs.id
+    }
+
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(id)
+    }
+}
