--- TVApp/TVApp/Services/JSONParser.swift
@@ -1,12 +1,0 @@
-//
-//  JSONParser.swift
-//  TVApp
-//
-//  Created by 지현우 on 2021/01/27.
-//
-
-import Foundation
-
-class JsonParser{
-    
-}
